requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cognitive-face==1.4.2
idna==2.7
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
